// src/app/core/models/masters/knowledge-hub/document-status.ts
export interface DocumentStatus {
    id: number;
    title: string;
    status: string;
    status_id :number;
    status_label: string;
    type: string;
}
export interface DocumentStatusPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: DocumentStatus[];
}